refactor: ByteUtils#bytesToMB

+++ commons/src/main/java/net/suqatri/cloud/commons/ByteUtils.java
@@ -2,8 +2,8 @@ package net.suqatri.cloud.commons;
 
 public class ByteUtils {
 
-    public static double bytesToMb(long bytes) {
-        return bytes / 1048576.0;
+    public static long bytesToMb(long bytes) {
+        return bytes / 1048576;
     }
 
 }
